+++ src/print.c
@@ -19,7 +19,6 @@ the Free Software Foundation, 675 Mass Ave, Cambridge, MA 02139, USA.  */
 
 
 #include <stdio.h>
-#include <string.h>
 #include "system.h"
 #include "machine.h"
 #include "new.h"
@@ -249,7 +248,7 @@ print_grammar()
   int i, j;
   short* rule;
   char buffer[90];
-  size_t column = 0;
+  int column = 0;
 
   /* rule # : LHS -> RHS */
   fputs("\nGrammar\n", foutput);
